added token check for user controller

+++ warlock-zone-spring/src/main/java/com/revature/warlockzone/controller/UserController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.warlockzone.beans.User;
@@ -23,6 +24,16 @@ public class UserController {
 	@Autowired
 	UserService userService;
 	
+	public boolean tokenCheck(String token) {
+		List<User> users = userService.getAllUsers();
+		for(int i = 0; i < users.size(); i++) {
+			if(users.get(i).getToken().equals(token)) {
+				return true;
+			}
+		}
+		
+		return false;
+	}
 	@RequestMapping(method = RequestMethod.POST, value = "/user")
 	public void addUser(@RequestBody User user) {
 		log.info("addUser " + user.toString());
@@ -36,9 +47,13 @@ public class UserController {
 	}
 	
 	@RequestMapping("/user")
-	public List<User> getUsers() {
-		log.info("getAllUsers");
-		return userService.getAllUsers();
+	public List<User> getUsers(@RequestParam("token") String tokens) {
+		if(tokenCheck(tokens)) {
+			log.info("getAllUsers");
+			return userService.getAllUsers();
+		
+		}
+		return null;
 	}
 	
 //	@RequestMapping("/user/{id}")
@@ -47,20 +62,29 @@ public class UserController {
 //	}
 	
 	@RequestMapping("/user/{username}")
-	public User getUserByUsername(@PathVariable String username) {
-		log.info("getUserByUsername: " + username);
-		return userService.getUserByUsername(username);
+	public User getUserByUsername(@PathVariable String username, @RequestParam("token") String tokens) {
+		if(tokenCheck(tokens)) {
+			log.info("getUserByUsername: " + username);
+			return userService.getUserByUsername(username);
+		}
+		return null;
 	}
 	
 	@RequestMapping(method = RequestMethod.PUT, value = "/user")
-	public void updateUser(@RequestBody User user) {
-		log.info("updateUser: " + user.toString());
-		userService.updateUser(user);
+	public void updateUser(@RequestBody User user, @RequestParam("token") String tokens) {
+		if(tokenCheck(tokens)) {
+			log.info("updateUser: " + user.toString());
+			userService.updateUser(user);
+			
+		}
 	}
 	
 	@RequestMapping(method = RequestMethod.DELETE, value = "/user/{id}")
-	public void deleteUserById(@PathVariable int id) {
-		log.warn("deleteUser: " + id);
-		userService.deleteUserById(id);
+	public void deleteUserById(@PathVariable int id , @RequestParam("token") String tokens) {
+		if(tokenCheck(tokens)) {
+			log.warn("deleteUser: " + id);
+			userService.deleteUserById(id);
+			
+		}
 	}
 }
